requirements.txt
# ── Core ──────────────────────────────────────────────────────
transformers>=5.1
torch
numpy
Pillow
requests
python-dotenv

# ── Depth Anything 3 ─────────────────────────────────────────
# Install from local checkout: pip install -e ./Depth-Anything-3
# depth-anything-3
